% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_splitcolumns.R
\name{build_splitcolumns}
\alias{build_splitcolumns}
\title{Returns a data frame 
with extra columns : LIB, VER, CASETYPE, FISS, FORM, SPEC}
\usage{
build_splitcolumns(df)
}
\arguments{
\item{df}{dataframe}
}
\description{
Returns a data frame 
with extra columns : LIB, VER, CASETYPE, FISS, FORM, SPEC
}
\examples{



}
