% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs-data.R
\docType{data}
\name{calcs}
\alias{calcs}
\title{A dataset of calcs ...}
\format{A data frame with 13282 observations of 7 variables:
\describe{
  \item{FULLID}{price, in US dollars}
  \item{MODEL}{weight of the diamond, in carats}
  \item{CODE}{weight of the diamond, in carats}
  \item{INST}{weight of the diamond, in carats}
  \item{LIBVER}{weight of the diamond, in carats}
  \item{CALCVAL}{weight of the diamond, in carats}
  \item{CALCERR}{weight of the diamond, in carats}
}}
\source{
\url{www.sendis.org}
}
\usage{
calcs
}
\description{
A dataset of calcs ...
}
\keyword{datasets}
