% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_format_errors.R
\name{check_format_errors}
\alias{check_format_errors}
\title{Checks that csv data is in expected format}
\usage{
check_format_errors(file, type = "df")
}
\arguments{
\item{file}{name of csv or dataframe to check}

\item{type}{type of file to check either 'csv' or 'df'}
}
\description{
Checks that csv data is in expected format
}
\examples{



}
