% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_to_sendis.R
\name{compile_to_sendis}
\alias{compile_to_sendis}
\title{Compiles dataframe to sendis format}
\usage{
compile_to_sendis(df)
}
\arguments{
\item{df}{dataframe}
}
\description{
This function is responsible for adding extra columns to the original 7 variable 
dataframe of calculations passed in argument so that it is in the sendis format (24 variables).
}
\details{
Warning : using this function will modify the global variable sendis for the session.
}
