% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exps-data.R
\docType{data}
\name{exps}
\alias{exps}
\title{A dataset of exps ...}
\format{A data frame with 2527 observations of 36 variables:
\describe{
  \item{FULLID}{price, in US dollars}
  \item{MODEL}{weight of the diamond, in carats}
  \item{CASETYPE}{weight of the diamond, in carats}
  \item{FISS}{weight of the diamond, in carats}
  \item{FORM}{weight of the diamond, in carats}
  \item{SPEC}{weight of the diamond, in carats}
  \item{EXPVAL}{weight of the diamond, in carats}
  \item{EXPERR}{weight of the diamond, in carats}
  \item{EALF}{weight of the diamond, in carats}
  \item{AFGE}{weight of the diamond, in carats} 
}}
\source{
\url{www.sendis.org}
}
\usage{
exps
}
\description{
A dataset of exps ...
}
\keyword{datasets}
