% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_submissions.R
\name{import_submissions}
\alias{import_submissions}
\title{Imports all csv files}
\usage{
import_submissions(path = "submissions")
}
\arguments{
\item{path}{name of directory with csv files to import. 
Defaults to directory 'submissions'.}
}
\description{
This function imports all csv format files inside specified directory in `path`. 
Returns a single data frame of 7 variables with all data, provided the resulting dataframe passes 
`check_format_errors()`.
}
