% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_userdata.R
\name{import_userdata}
\alias{import_userdata}
\title{Import csv data file from user and returns a data frame 
columns CASETYPE, FISS, FORM, SPEC are reconstructed from column FULLID}
\usage{
import_userdata(filename)
}
\arguments{
\item{filename}{filename in csv format to import.}
}
\description{
Import csv data file from user and returns a data frame 
columns CASETYPE, FISS, FORM, SPEC are reconstructed from column FULLID
}
\examples{



}
