% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mats-data.R
\docType{data}
\name{mats}
\alias{mats}
\title{A dataset containing Z, A (isotope list) data}
\format{A data frame with 562 observations of 5 features:
\describe{
  \item{MAT}{MAT number in ENDF-6 format}
  \item{Z}{Number of protons}
  \item{X}{Isotope symbol}
  \item{A}{Atomic mass}
  \item{M}{Isomer state}
}}
\usage{
mats
}
\description{
This is an export of the isotope list of JEFF-3.3. It contains Z,A pairs and mat numbers
for  562 Isotopes.
}
\keyword{datasets}
