% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cumulchi.R
\name{plot_cumulchi}
\alias{plot_cumulchi}
\title{plot cumulative Chi-square function}
\usage{
plot_cumulchi(df)
}
\arguments{
\item{df}{dataframe to consider (no default)}
}
\description{
plot cumulative Chi-square function
}
