% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ealf.R
\name{plot_ealf}
\alias{plot_ealf}
\title{Plot C over E vs EALF.}
\usage{
plot_ealf(df)
}
\arguments{
\item{df}{dataframe to consider (no default),}
}
\description{
Plot C over E vs EALF.
}
\examples{

plot_ealf(sendis)

}
