% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hcluster.R
\name{plot_hc}
\alias{plot_hc}
\title{Plot hierarchical clustering}
\usage{
plot_hc(df, dmethod = "euclidean", clustering = "single", k)
}
\arguments{
\item{df}{dataframe to consider (no default),}
}
\description{
Plot hierarchical clustering
}
