% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_table_header.R
\name{rename_table_header}
\alias{rename_table_header}
\title{Restyle header for sendis-report.}
\usage{
rename_table_header(df)
}
\arguments{
\item{df}{dataframe to consider (no default),}
}
\description{
Restyle header for sendis-report.
}
