% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_chitable.R
\name{return_chitable}
\alias{return_chitable}
\title{Returns dataframe with total chi-square values}
\usage{
return_chitable(df)
}
\arguments{
\item{df}{dataframe to consider (no default),}
}
\description{
Returns dataframe with total chi-square values
}
