% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_common_suite.R
\name{return_common_suite}
\alias{return_common_suite}
\title{Returns dataframe df with only common cases specified by inst}
\usage{
return_common_suite(df, inst)
}
\arguments{
\item{df}{dataframe to consider (no default),}
}
\description{
Returns dataframe df with only common cases specified by inst
}
