% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_discrepant.R
\name{return_discrepant}
\alias{return_discrepant}
\title{Return a dataframe of common benchmarks}
\usage{
return_discrepant(df, INST1, INST2, sigma_treshold = 0)
}
\arguments{
\item{df}{dataframe to explore (no default).}

\item{INST1}{(no default),}

\item{INST2}{(no default),}

\item{sigma_treshold}{(no default),}
}
\description{
And filters them according to their difference.
}
