% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sendis}
\alias{sendis}
\title{A dataset containing sendis data etc ...}
\format{A data frame with 13278 observations of 15 variables:
\describe{
  \item{FULLID}{price, in US dollars}
  \item{MODEL}{weight of the diamond, in carats}
  \item{CASETYPE}{weight of the diamond, in carats}
  \item{FISS}{weight of the diamond, in carats}
  \item{FORM}{weight of the diamond, in carats}
  \item{SPEC}{weight of the diamond, in carats}
  \item{EXPVAL}{weight of the diamond, in carats}
  \item{EXPERR}{weight of the diamond, in carats}
  \item{EALF}{weight of the diamond, in carats}
  \item{AFGE}{weight of the diamond, in carats}
  \item{CODE}{weight of the diamond, in carats}
  \item{INST}{weight of the diamond, in carats}
  \item{LIBVER}{weight of the diamond, in carats}
  \item{CALCVAL}{weight of the diamond, in carats}
  \item{CALCERR}{weight of the diamond, in carats}
}}
\source{
\url{www.sendis.org}
}
\usage{
sendis
}
\description{
A dataset containing sendis data etc ...
}
\keyword{datasets}
