% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens-data.R
\docType{data}
\name{sens}
\alias{sens}
\title{A dataset containing sensitivity data etc ...}
\format{A data frame with 236128 observations of 7variables:
\describe{
  \item{FULLID}{price, in US dollars}
  \item{ISOTOPE}{weight of the diamond, in carats}
  \item{KSENS1}{weight of the diamond, in carats}
  \item{KSENS2}{weight of the diamond, in carats}
  \item{KSENS3}{weight of the diamond, in carats}
  \item{KSENSTOT}{weight of the diamond, in carats}
  \item{REACTION}{weight of the diamond, in carats}
}}
\source{
\url{www.sendis.org}
}
\usage{
sens
}
\description{
A dataset containing sensitivity data etc ...
}
\keyword{datasets}
