% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_derived_calcs.R
\name{build_derived_calcs}
\alias{build_derived_calcs}
\title{Computes and adds columns :}
\usage{
build_derived_calcs(df = calcs)
}
\arguments{
\item{df}{dataframe to consider (no default),}
}
\description{
Columnds added are : 
            
      * EXPVAL
      * EXPERR
      * COVERE
      * RESIDUAL
      * CUMUL
      * CHISQ
}
\examples{
\dontrun{
library(sendis)

build_derived_calcs(calcs)

}

}
