% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scree.R
\name{plot_scree}
\alias{plot_scree}
\title{plot scree plot for clustering analysis}
\usage{
plot_scree(df, cmax = 20, nstart = 20)
}
\arguments{
\item{df}{dataframe to consider (no default)}
}
\description{
plot scree plot for clustering analysis
}
